Move PubSub from Producer namespace
